refactor (transform: move db methods to a separate file
